+++ examples/__tests__/integration/nitro.test.ts
@@ -1,0 +1,5 @@
+import { deployExample } from '../test-utils';
+// eslint-disable-next-line jest/no-disabled-tests
+it.skip('[examples] should deploy nitro', async () => {
+  await deployExample('nitro');
+});