+++ examples/nitro/server/routes/index.ts
@@ -1,0 +1,3 @@
+export default defineEventHandler((event) => {
+  return "Start by editing <code>server/routes/index.ts</code>.";
+});