+++ packages/cli/src/util/telemetry/commands/cache/index.ts
@@ -1,0 +1,15 @@
+import { TelemetryClient } from '../..';
+import type { TelemetryMethods } from '../../types';
+import type { cacheCommand } from '../../../../commands/cache/command';
+
+export class CacheTelemetryClient
+  extends TelemetryClient
+  implements TelemetryMethods<typeof cacheCommand>
+{
+  trackCliSubcommandPurge(actual: string) {
+    this.trackCliSubcommand({
+      subcommand: 'purge',
+      value: actual,
+    });
+  }
+}